Audit and fix correctness bugs in CP code sections
Per-section audit against KACTL, CP-Algorithms, AtCoder Library and
cppreference found ~17 correctness bugs that would cause WA/TLE/RE/MLE
in a contest. Fixed in 18 files. Changes were conservative — only
correctness, no style or refactoring.

Notable fixes:

- Kruskal.cpp: replaced broken sup[] cycle-tracker with proper DSU
- Bridges.cpp: bridge condition was inverted (disc<low instead of low>disc)
- IterativeSegTree.cpp: build() was initializing leaves to 1 instead of a[i]
- FFT.cpp: removed hardcoded MAX=2^18, switched to dynamic sizing
- EulerTotientFunction.cpp: removed duplicate function definitions
- MatrixDeterminant.cpp / SieveOfEratosthenes.cpp: converted to proper
  functions taking parameters (originals used uninitialized global n)
- ManacherAlgorithm.cpp: replaced stub with full manacher_odd + manacher_even
- DPOnTrie.cpp: missing MOD constant
- XORBasis.cpp: removed duplicate function block (multiple-definition linker error)
- AllSubsets.cpp: undefined min_diff + loop bound off-by-one
- BerlekampMassey.cpp: wrong polynomial initialization on first failure
- DigitDP.cpp: dp array now lazily initialized (note: single-test-case only)
- CoinProblem.cpp: INT_MAX -> LLONG_MAX for type consistency with vector<ll>
- LineSegmentIntersect.cpp: added collinear overlap check
- CentroidDecomposition.cpp: added dst[0]=1 marker after fill
- combinatorics_calc.py: i_fact[0] was never computed (off-by-one in range)
- test_graph.py: tree case was missing edge count on first line

All changes verified against reference implementations; no test cases
were run (no fixtures available). Recommend compile-check + spot test
before pushing.

LaTeX compile was not verified here (no pdflatex available).

+++ code/DynamicProgramming/CoinProblem.cpp
@@ -5,7 +5,7 @@ void solve(){
         cin>>coins[i];
     }
     
-    vector<ll> dp(x+1,INT_MAX);
+    vector<ll> dp(x+1,LLONG_MAX);
     dp[0] = 0;
     for(int i=0; i<=x; i++){
         for(int j=0; j<n; j++){
@@ -16,7 +16,7 @@ void solve(){
         }
     }
     
-    if(dp[x] != INT_MAX){
+    if(dp[x] != LLONG_MAX){
         cout<<dp[x]<<endl;
     }else{
         cout<<"-1"<<endl;
